+++ cx2t/src/claw/tatsu/xcodeml/xnode/fortran/FortranType.java
@@ -6,7 +6,6 @@ package claw.tatsu.xcodeml.xnode.fortran;
 
 import claw.tatsu.xcodeml.xnode.Xname;
 
-import java.security.SecureRandom;
 import java.util.Random;
 
 /**