+++ test/claw/sca/sca14/reference_cpu.f90
@@ -16,8 +16,6 @@ CONTAINS
   DO k = 2 , nz , 1
    DO proma = 1 , nproma , 1
     t ( proma , k ) = c * k
-   END DO
-   DO proma = 1 , nproma , 1
     q ( k , proma , 1 ) = q ( k - 1 , proma , 1 ) + t ( proma , k ) * c
    END DO
   END DO